New Feature: Supporting application basePath
In order to support the same application for different environment on the same URL, we need to consider the window.location.basepath varible.

Introduced a changed to the library to support better testability. If you do nothing it's business as usual but if you need to test specifc things related to window.location, you can override the default value.